backend/requirements.txt
numpy>=2.0,<2.3
opencv-python==4.12.0.88
scikit-image==0.25.0
scikit-learn==1.6.1
tensorflow>=2.18.0
Pillow==10.4.0
fastapi==0.116.1
uvicorn==0.34.0
python-multipart==0.0.19
matplotlib==3.10.5
tqdm==4.66.5

-e .
